+++ docs/source/change_log.rst
@@ -10,6 +10,7 @@ API
 ---
 * Adds `add_attendees` method to the `Event` for adding multiple attendees.
 * Add specific time reminders (N days before at HH:MM)
+* Support Python3.12
 
 Core
 ----